01-Projects/03-ML/Facebook/requirements.txt
facebook-api
requests
scikit-learn
matplotlib
textblob
Arabic-Stopwords
Tashaphyne